phase 5: modularize the repair-screen-ios.tsx code into hooks and components - 1000 to 300 LOC
